Move upload middleware too programmatic middleware
