--- packages/core/upload/server/middlewares/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const upload = require('./upload');
-
-module.exports = {
-  upload,
-};
