+++ packages/core/upload/server/register.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const registerUploadMiddlware = require('./controllers/upload-middleware');
+
+module.exports = async () => {
+  await registerUploadMiddlware();
+};
